% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\name{Example}
\alias{Example}
\title{Writes QAQC report in html format for example data provided with package.}
\usage{
Example(user_data_folder = NULL)
}
\arguments{
\item{user_data_folder}{file path to user data folder, defaults to Desktop/LinkrealTimeloads_Output}
}
\value{
List with suspended-sediment load estimates at Johnstone River at Innisfail saved to user_data_folder/site/Loads.rds
}
\description{
Writes folders and files using code in vignette('LinkrealTimeloads') to folder path specified by user_data_folder. Uses example data files provided package folder LinkrealTimeloads/extdata. The QAQC report in html format provides interactive and shareable plots. Example data provided from the Department of Environment and Science Great Barrier Reef Catchment Loads Monitoring Program (GBRCLMP). Please note no attempt has been made to remove data affected by sensor or sampling malfunctions, data are provided for illustration purposes only.
}
\note{
All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <-
"JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\examples{
\dontrun{
# Write to Desktop/LinkrealTimeloads_Output
Output <- Example()
}
}
\references{
Example data provided from the Department of Environment and Science Great Barrier Reef Catchment Loads Monitoring Program (GBRCLMP). Please note no attempt has been made to remove data affected by sensor or sampling malfunctions, data are provided for illustration purposes only.

Stephen Wallace (2023, DES) provided crucial functions in Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of Concentration and Discharge. R package version 1.0.0.

Livsey, D.N. (in review). National Industry Guidelines for hydrometric monitoring–Part 12: Application of acoustic Doppler velocity meters to measure suspended-sediment load. Bureau of Meteorology. Melbourne, Australia
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
