% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Channelmaster_Data.R
\name{Import_Channelmaster_Data}
\alias{Import_Channelmaster_Data}
\title{Imports data from binary RDI 000 files using code from Stephen Wallace and moves variables to format expected in the package realTimeloads}
\usage{
Import_Channelmaster_Data(user_data_folder, site = NULL, overwrite = FALSE)
}
\arguments{
\item{user_data_folder}{file path to user data folder}

\item{site}{site folder under user_data_folder}

\item{overwrite}{overwrite all previously imported data}
}
\value{
A rds file for each 000 file and All_ADCP_Programming_Data.rds
}
\description{
Imports data from binary RDI 000 files when new files are added OR when 000 file size changes. Stephen Wallace provided all functions that extract data from 000 files. This function we needed to assign variable names expected in realTimeloads.
}
\note{
All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <- "JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\examples{
# See vignette('LinkrealTimeloads',package = LinkrealTimeloads)
}
\references{
Stephen Wallace (2023, DES) provided crucial functions in Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of Concentration and Discharge_. R package version 1.0.0.
}
\seealso{
\code{\link{Link_to_Real_time_loads}} Process data in specified folder structure using realTimeloads package
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
