% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Impute_height_discharge_velocity_and_conductivity.R
\name{Impute_height_discharge_velocity_and_conductivity}
\alias{Impute_height_discharge_velocity_and_conductivity}
\title{Infills missing height, discharge, and conductivity data using realTimeloads::impute_data()}
\usage{
Impute_height_discharge_velocity_and_conductivity(
  user_data_folder,
  site,
  from_time,
  to_time,
  overwrite = FALSE,
  MC = 1
)
}
\arguments{
\item{user_data_folder}{file path to user data folder}

\item{site}{site folder under user_data_folder}

\item{from_time}{date to start imputation}

\item{to_time}{date to end imputation}

\item{overwrite}{logical to overwrite all previous imputation data}

\item{MC}{number of Monte Carlo simulations used in uncertainty estimation in realTimeloads::impute_data}
}
\value{
Imputed data in rds files *_Imputed.rds
}
\description{
Code will impute data if days since last imputation threshold since "to_time" is reached
}
\note{
All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <- "JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\section{Warning}{

Users should ensure imputation is reasonable and adjust predictors as needed.
}

\examples{
# See vignette('LinkrealTimeloads',package = LinkrealTimeloads)
}
\references{
Stephen Wallace (2023, DES) provided crucial functions in Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of Concentration and Discharge_. R package version 1.0.0.
}
\seealso{
\code{\link{Link_to_Real_time_loads}} Process data in specified folder structure using realTimeloads package

\code{\link{Predict_TSS_from_SCI}} Estimates TSS from processed acoustic backscatter and turbidity
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
