% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_TSS.R
\name{Impute_TSS}
\alias{Impute_TSS}
\title{Infills missing TSS data using realTimeloads::impute_data()}
\usage{
Impute_TSS(user_data_folder, site = NULL, MC = 1, ptrain = 0.9)
}
\arguments{
\item{user_data_folder}{file path to user data folder}

\item{site}{site folder under user_data_folder}

\item{MC}{number of Monte Carlo simulations used in uncertainty estimation in realTimeloads::impute_data}

\item{ptrain}{proportion of data to use in training of decision trees in realTimeloads::impute_data()}
}
\value{
Imputed data in /Analyte_data/TSS_Imputed.rds
}
\description{
Imputes TSS using tidal model fit in realTimeloads::impute_data(). I tried the Froude number and tidally filtered discharge as predictors but tidal model predictors provided reasonable fits. In https://aha.net.au/wp-content/uploads/AHA-Journal-Spring-2022_HR.pdf I show TSS imputed using the Froude number, tidally filtered discharge , and a nonlinear auto regressive exogenous model (NARX) model. A NARX model is similar to the Bidirectional RNN in keras see https://keras.io/guides/working_with_rnns/
}
\note{
All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <- "JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\section{Warning}{

Users should ensure imputation is reasonable and adjust predictors as needed.
}

\examples{
# See vignette('LinkrealTimeloads',package = LinkrealTimeloads)
}
\references{
Stephen Wallace (2023, DES) provided crucial functions in Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of Concentration and Discharge_. R package version 1.0.0.
}
\seealso{
\code{\link{Link_to_Real_time_loads}} Process data in specified folder structure using realTimeloads package
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
