% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_to_Real_time_loads.R
\name{Link_to_Real_time_loads}
\alias{Link_to_Real_time_loads}
\title{Import and process data in specified folder structure using realTimeloads package}
\usage{
Link_to_Real_time_loads(user_data_folder, site = NULL, overwrite = FALSE)
}
\arguments{
\item{user_data_folder}{file path to user data folder}

\item{site}{site folder(s) under user_data_folder}

\item{overwrite}{logical to overwrite all previous outputs}
}
\value{
Processed data in subfolders of user_data_folder
}
\description{
Imports, processes, and imputes data in user_data_folder folders (see note below). Calling this function periodically will check for changes to 000 file size and will appended new data as needed. After running this function use Compute_load() (see link below)
}
\note{
All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <- "JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\section{Warning}{

Only set overwrite = T if you understand what will happen in each function below

See note above on file structure

Site_list_update() and Site__and_Channel_Information_Update() must be manually updated when adding new sites
}

\examples{
# See vignette('LinkrealTimeloads',package = LinkrealTimeloads)
}
\references{
Stephen Wallace (2023, DES) provided crucial functions in Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of Concentration and Discharge_. R package version 1.0.0.

Livsey, D.N. (in review). National Industry Guidelines for hydrometric monitoring–Part 12: Application of acoustic Doppler velocity meters to measure suspended-sediment load. Bureau of Meteorology. Melbourne, Australia
}
\seealso{
\code{\link{Compute_load}} Computes load generated by outputs from Link_to_Real_time_loads()
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
