% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkrealTimeloads-package.R
\docType{package}
\name{LinkrealTimeloads-package}
\alias{LinkrealTimeloads}
\alias{LinkrealTimeloads-package}
\title{LinkrealTimeloads: Code to process data from DES WQI GBRCLMP using the R package "realTimeloads"}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel Livsey \email{livsey.daniel@gmail.com} (\href{https://orcid.org/0000-0002-2028-6128}{ORCID}) [copyright holder]

}
\keyword{internal}
