% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_f20.R
\name{predict_f20}
\alias{predict_f20}
\title{Predict the fraction of TSS/SSC that is less than 20 um by mass using water velocity, turbidity, and salinity}
\usage{
predict_f20(
  Water_velocity_in_m_per_s,
  Total_depth_in_m,
  Salinity_in_PSU,
  Turbidity_in_FNU,
  Sample_depth_in_m = NULL,
  ptrain = 0.8,
  MC = 1
)
}
\arguments{
\item{Water_velocity_in_m_per_s}{cross-section stream velocity (m/s)}

\item{Total_depth_in_m}{stream depth (m)}

\item{Salinity_in_PSU}{Salinity (PSU)}

\item{Turbidity_in_FNU}{Turbidity (FNU)}

\item{Sample_depth_in_m}{Depth of water sample or monitoring point (m)}

\item{ptrain}{Fraction of particle size training data to use (-)}

\item{MC}{Number of Monte Carlo simulations (-)}
}
\value{
estimates of the fraction of sediment less than 20 um by mass
}
\description{
Predict the fraction of TSS/SSC that is less than 20 um by mass using decision tree algorithms following Livsey et al (2022). Uncertainty on predictions is estimated using methods adopted from Rustomji and Wilkinson (2008). When multiplied by TSS, outputs can be used to estimate the total load that is less than or greater than 20 um
}
\note{
Conductivity can be converted into Salinity in PSU using realTimeloads::ctd2sal()
}
\examples{
#  idealized gradient in turbidity offshore
n <- 10
U <- rep(1,n)
H <- rep(10,n)
PSU <- pracma::linspace(0,32,n)
FNU <- pracma::linspace(50,1,n)
SD <- 0.33*H
f20 = predict_f20(U,H,PSU,FNU,SD)
}
\references{
Rustomji, P., & Wilkinson, S. N. (2008). Applying bootstrap resampling to quantify uncertainty in fluvial suspended sediment loads estimated using rating curves. Water resources research, 44(9).https://doi.org/10.1029/2007WR006088

Livsey, D. N., Crosswell, J. R., Turner, R. D. R., Steven, A. D. L., & Grace, P. R. (2022). Flocculation of Riverine Sediment Draining to the Great Barrier Reef, Implications for Monitoring and Modeling of Sediment Dispersal Across Continental Shelves. Journal of Geophysical Research: Oceans, 127(7), https://doi.org/10.1029/2021JC017988
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
