% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_TSS_from_SCI.R
\name{Predict_TSS_from_SCI}
\alias{Predict_TSS_from_SCI}
\title{Predict TSS from processed acoustic backscatter and turbidity using Livsey et al (2023)}
\usage{
Predict_TSS_from_SCI(
  user_data_folder,
  site = NULL,
  overwrite = FALSE,
  retain_data_for_load_uncertainty = FALSE
)
}
\arguments{
\item{user_data_folder}{file path to user data folder}

\item{site}{site folder under user_data_folder}

\item{overwrite}{logical to overwrite all previous TSS estimates}

\item{retain_data_for_load_uncertainty}{logical to keep estimates of Kt for later load uncertainty calculations}
}
\value{
TSS_in_mg_per_liter_estimated_from_SCI in rds files
}
\description{
Code expects outputs from Link_to_Real_time_loads. Conducts regression of TSS(backscatter,turbidity) using realTimeloads package code.
}
\note{
When retain_data_for_load_uncertainty = TRUE, exported rds file can be very large

All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <- "JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\section{Warning}{

Link_to_Real_time_loads should be run first

When retain_data_for_load_uncertainty = TRUE, exported rds file can be very large.
For load uncertainty calculations one must store 2000 Monte Carlo simulations per
time step (see realTimeloads documentation and references therein)
}

\examples{
# See vignette('LinkrealTimeloads',package = LinkrealTimeloads)
}
\references{
Stephen Wallace (2023, DES) provided crucial functions in Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of Concentration and Discharge_. R package version 1.0.0.

Livsey, D. N., Turner, R. D. R., & Grace, P. R. (2023). Combining Optical and Acoustic Backscatter Measurements for Monitoring of Fine Suspended‐Sediment Concentration Under Changes in Particle Size and Density. Water Resources Research, 59(8), e2022WR033982.

Livsey, D.N. (in review). National Industry Guidelines for hydrometric monitoring–Part 12: Application of acoustic Doppler velocity meters to measure suspended-sediment load. Bureau of Meteorology. Melbourne, Australia
}
\seealso{
\code{\link{Link_to_Real_time_loads}} Process data in specified folder structure using realTimeloads package
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
