% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Backscatter.R
\name{Process_Backscatter}
\alias{Process_Backscatter}
\title{Process acoustic backscatter and save to ADCP (i.e., 000) rds file}
\usage{
Process_Backscatter(user_data_folder, site = NULL, overwrite = FALSE)
}
\arguments{
\item{user_data_folder}{file path to user data folder}

\item{site}{name of site folder under user_data_folder}

\item{overwrite}{overwrite previous outputs}
}
\value{
Saves processed acoustic backscatter to each 000-related rds file in site/ADCP_data

\code{\link{Link_to_Real_time_loads}} Process data in specified folder structure using realTimeloads package
\code{\link{Import_Channelmaster_Data}} Imports ADCP data from 000 files
}
\description{
Process acoustic backscatter using realTimeloads with methods detailed in Livsey (in review)
}
\note{
All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <- "JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\section{Warning}{

Import_Channelmaster_Data() should be run first
}

\examples{
# See vignette('LinkrealTimeloads',package = LinkrealTimeloads)
}
\references{
Stephen Wallace (2023, DES) provided crucial functions in Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of Concentration and Discharge_. R package version 1.0.0.

Livsey, D.N. (in review). National Industry Guidelines for hydrometric monitoring–Part 12: Application of acoustic Doppler velocity meters to measure suspended-sediment load. Bureau of Meteorology. Melbourne, Australia
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
