% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_data_and_compute_load.R
\name{Process_data_and_compute_load}
\alias{Process_data_and_compute_load}
\title{Processes all data in site folders and compute load}
\usage{
Process_data_and_compute_load(
  user_data_folder,
  site,
  Site_Number = NULL,
  Site_Name = NULL,
  wait_for_user = FALSE,
  write_QAQC_report = FALSE,
  QAQC_Parameters = NULL
)
}
\arguments{
\item{user_data_folder}{file path to user data folder}

\item{site}{site folder under user_data_folder}

\item{Site_Number}{site number for site}

\item{Site_Name}{site name for plotting purposes}

\item{wait_for_user}{Boolean to wait for user to place files}

\item{write_QAQC_report}{Boolean to call QAQC_Report()}

\item{QAQC_Parameters}{list of parameters needed for QAQC_Report()}
}
\value{
List with suspended-sediment load estimates and saved to user_data_folder/site/Loads.rds
}
\description{
Processes data in site folders and computes load after user has manually placed files in site folder. If wait_for_user = TRUE initialize_new_site_directories() writes directories expected by package code and provides dialogue box to wait for user to indicate when data files have been placed into site folders
}
\note{
All functions expect data to be in following folders:

\describe{
\item{1. user_data_folder/site/ADCP_data}{- Place 000 files here}
}
\describe{
\item{2. user_data_folder/site/Analyte_data}{- Place TSS data from EAGLE IO csv export here}
}
\describe{
\item{3. user_data_folder/site/Channel_Geometry}{- Place cross-section survety in xs.RDS file format from Stephen Wallace here}
}
\describe{
\item{4. user_data_folder/site/Discharge_data }{- Place discharge and velocity csv files from Stephen Wallace here}
}
\describe{
\item{5. user_data_folder/site/Height_Offsets}{- Place height offset data from EAGLE IO csv export here}
}
\describe{
\item{6. user_data_folder/site/Sonde_and_Height_data}{- Place TSS data from EAGLE IO csv export here}
}

Where "site" is a specific site (e.g., for Johnstone at Innisfail use site <-
"JRI")

See vignette('LinkrealTimeloads') for further instructions
}
\examples{
\dontrun{
Output <- Process_data_and_compute_load(user_data_folder,site,wait_for_user = FALSE)
}
}
\references{
Stephen Wallace (2023, DES) provided crucial functions in
Import_Channelmaster_Data() to extract data from binary 000 files.

Livsey D.N. (2023). realTimeloads: Analyte Flux and Load from Estimates of
Concentration and Discharge_. R package version 1.0.0.

Livsey, D.N. (in review). National Industry Guidelines for hydrometric
monitoring–Part 12: Application of acoustic Doppler velocity meters to
measure suspended-sediment load. Bureau of Meteorology. Melbourne, Australia
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
